refactor: move tracking logic from lite core to lite-ui
lite core now exports only `setControllerReadHook` — a single callback
hook that fires on every `ctrl.get()`. All tracking machinery (track,
registerInTracker, startTracking, startArrayTracking) moved to
`packages/lite-ui/src/tracking.ts`. lite-ui installs the hook at import
time via `setControllerReadHook()`.

This keeps the DI/effect core clean of UI reactivity concerns.
