+++ packages/lite-ui/src/react.ts
@@ -2,7 +2,7 @@ import type { ComponentType } from 'react'
 import { createElement } from 'react'
 import { createRoot, type Root } from 'react-dom/client'
 import { ScopeProvider } from '@pumped-fn/lite-react'
-import { track } from '@pumped-fn/lite'
+import { track } from './tracking'
 import { DIRECTIVE_BRAND, subscribeToControllers, type Directive, type MountContext, type ReactiveBinding } from './index'
 
 export function react<P extends Record<string, unknown>>(
